+++ apps/comparative/double-slit-analytical/js/release.js
@@ -1,3 +1,1 @@
-export const APP_RELEASE = Object.freeze({version:74,date:'2026-09-19'});
-
-
+export const APP_RELEASE = Object.freeze({version:75,date:'2026-09-19'});